Multi-step register and update form
