+++ src/app/form-controls/datastream/datastream.component.scss
@@ -1,3 +1,3 @@
-@import "../../forms/sensor-register/sensor-register.component.scss";
+@import "../../forms/sensor/sensor.component";
