+++ src/app/form-controls/sensor-theme/sensor-theme.component.scss
@@ -1,1 +1,1 @@
-@import "../../forms/sensor-register/sensor-register.component.scss";
+@import "../../forms/sensor/sensor.component";
