+++ src/app/form-controls/sensor-type/sensor-type.component.scss
@@ -1,1 +1,1 @@
-@import "../../forms/sensor-register/sensor-register.component.scss";
+@import "../../forms/sensor/sensor.component";
