+++ src/app/form-controls/sensor-type/sensor-type.component.ts
@@ -104,7 +104,9 @@ export class SensorTypeComponent implements ControlValueAccessor, OnDestroy {
     });
 
     this.form.get('typeDetails').valueChanges.subscribe((category: Category) => {
-      this.sensorType.emit(this.form.get('typeDetails').value);
+      if (category) {
+        this.sensorType.emit(category);
+      }
     });
   }
