+++ src/app/forms/owner-update/owner-update.component.html
@@ -49,7 +49,6 @@
                 </div>
                 <hr>
                 <button class="btn btn-primary btn-lg mr-3 mb-1" type="submit">Update</button>
-                <button type="button" class="btn btn-lg mb-1" (click)="close()">Close</button>
             </form>
         </div>
     </div>
