+++ src/app/forms/owner-update/owner-update.component.scss
@@ -1,4 +1,4 @@
-@import "../sensor-register/sensor-register.component";
+@import "../sensor/sensor.component";
 
 .form-header {
   margin-top: 72px;
