--- src/app/forms/sensor-register/sensor-register.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SensorRegisterComponent } from './sensor-register.component';
-
-describe('SensorRegisterComponent', () => {
-  let component: SensorRegisterComponent;
-  let fixture: ComponentFixture<SensorRegisterComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ SensorRegisterComponent ],
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(SensorRegisterComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  xit('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
