--- src/app/forms/sensor-update/sensor-update.component.scss
@@ -1,19 +1,0 @@
-label {
-  margin-top: 5px;
-  color: #000000;
-  font-style: normal;
-  font-weight: normal
-}
-
-.form-check-input {
-  margin-left: -1.25rem;
-  margin-right: 0.5em;
-}
-
-.form-check-label {
-  margin-left: 5px;
-}
-
-.form-header {
-  margin-top: 72px;
-}
