+++ src/app/forms/sensor/sensor.component.scss
