+++ src/app/forms/sensor/sensor.component.ts
@@ -1,0 +1,310 @@
+import { Subscription } from 'rxjs';
+import { ISensor } from '../../model/bodies/sensor-body';
+import { ActivatedRoute, Router } from '@angular/router';
+import { AlertService } from '../../services/alert.service';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { LocationService } from '../../services/location.service';
+import { FormGroup, Validators, FormBuilder, FormArray } from '@angular/forms';
+import { IRegisterSensorBody, IUpdateSensorBody, SensorService } from '../../services/sensor.service';
+
+@Component({
+  selector: 'app-sensor',
+  templateUrl: './sensor.component.html',
+  styleUrls: ['./sensor.component.scss'],
+})
+export class SensorComponent implements OnInit, OnDestroy {
+  public sensor: ISensor;
+
+  public submitted = false;
+  public activeStepIndex = 0;
+
+  public subscriptions: Subscription[] = [];
+
+  public form: FormGroup;
+  public formControlSteps: Record<string, Array<any>>;
+
+  public urlRegex = '(https?://)?([\\da-z.-]+)\\.([a-z.]{2,6})[/\\w .-]*/?';
+
+  constructor(
+    private readonly router: Router,
+    private readonly route: ActivatedRoute,
+    private readonly formBuilder: FormBuilder,
+    private readonly alertService: AlertService,
+    private readonly sensorService: SensorService,
+    private readonly locationService: LocationService,
+  ) {}
+
+  get f() {
+    return this.form.controls;
+  }
+
+  public setName(item: string) {
+    this.form.controls.name.setValue(item);
+  }
+
+  public goToStep(step: number): void {
+    if (Object.values(this.formControlSteps)[this.activeStepIndex].some((f) => f.invalid)) {
+      if (step <= this.activeStepIndex) {
+        this.activeStepIndex = step;
+      } else {
+        this.submitted = true;
+      }
+    } else {
+      this.submitted = false;
+      this.activeStepIndex = step;
+    }
+  }
+
+  public getStepCount(): number {
+    return Object.keys(this.formControlSteps).length;
+  }
+
+  public getStepLabel(): string {
+    return Object.keys(this.formControlSteps)[this.activeStepIndex];
+  }
+
+  public getStepClasses(pageIndex: number) {
+    return { active: this.activeStepIndex === pageIndex, finished: this.activeStepIndex > pageIndex };
+  }
+
+  public setSensor(sensor: ISensor): void {
+    this.sensor = sensor;
+
+    this.form.patchValue({
+      aim: this.sensor.aim || '',
+      description: this.sensor.description || '',
+      manufacturer: this.sensor.manufacturer || '',
+      active: { value: this.sensor.active.toString() || false },
+      documentationUrl: this.sensor.documentationUrl || '',
+      location: {
+        longitude: this.sensor.location ? this.sensor.location.coordinates[0] : null,
+        latitude: this.sensor.location ? this.sensor.location.coordinates[1] : null,
+        height: this.sensor.location ? this.sensor.location.coordinates[2] : null,
+        baseObjectId: this.sensor.baseObjectId || 'non-empty',
+      },
+      type: {
+        typeName: this.sensor.typeName ? this.sensor.typeName[0] : '',
+        typeDetails: this.sensor.typeDetails ? this.sensor.typeDetails.subType : '',
+      },
+      theme: { value: this.sensor.theme || [] },
+    });
+
+    // Overwrite the name generated by type.
+    this.form.patchValue({
+      name: this.sensor.name || '',
+    });
+
+    const dataStreams = this.form.get('dataStreams') as FormArray;
+    dataStreams.clear();
+    for (const dataStream of this.sensor.dataStreams) {
+      dataStreams.push(this.formBuilder.group({
+        dataStreamId: dataStream.dataStreamId,
+        name: [dataStream.name, Validators.required],
+        reason: dataStream.reason || '',
+        description: dataStream.description || '',
+        observedProperty: dataStream.observedProperty || '',
+        isPublic: !!dataStream.isPublic,
+        isOpenData: !!dataStream.isOpenData,
+        isReusable: !!dataStream.isReusable,
+        documentationUrl: [dataStream.documentationUrl || '', [Validators.pattern(this.urlRegex)]],
+        dataLink: [dataStream.dataLink || '', [Validators.pattern(this.urlRegex)]],
+        unitOfMeasurement: dataStream.unitOfMeasurement || '',
+        dataFrequency: dataStream.dataFrequency || 0,
+        dataQuality: dataStream.dataQuality || 0,
+      }));
+    }
+  }
+
+  public async submit() {
+    this.submitted = true;
+
+    // Stop if the form is invalid.
+    if (this.form.valid) {
+      if (this.sensor) {  // Update the sensor.
+        const updatedSensorProperties = {
+          typeName: this.form.value.type.typeName,
+          location: this.form.value.location,
+          dataStreams: this.form.value.dataStreams,
+          active: this.form.value.active || false,
+          aim: this.form.value.aim,
+          description: this.form.value.description !== '' ? this.form.value.description : undefined,
+          documentationUrl: this.form.value.documentationUrl !== '' ? this.form.value.documentationUrl : undefined,
+          manufacturer: this.form.value.manufacturer,
+          name: this.form.value.name,
+          theme: this.form.value.theme.value,
+          typeDetails: {subType: this.form.value.type.typeDetails},
+        };
+
+        try {
+          const active: boolean = JSON.parse(updatedSensorProperties.active.value);
+          if (active === true && !this.sensor.active) {
+            await this.sensorService.activate(this.sensor._id);
+          } else if (active === false && this.sensor.active) {
+            await this.sensorService.deactivate(this.sensor._id);
+          }
+        } catch (error) {
+          console.error(error);
+        }
+
+        try {
+          const newDataStreams = updatedSensorProperties.dataStreams || [];
+          const newDataStreamIds = newDataStreams.map((d) => d.dataStreamId);
+          for (let newDataStream of newDataStreams) {
+            if (!newDataStream.hasOwnProperty('dataStreamId')) {
+              newDataStream = {
+                name: newDataStream.name,
+                reason: newDataStream.reason || undefined,
+                description: newDataStream.description || undefined,
+                observedProperty: newDataStream.observedProperty || undefined,
+                isPublic: newDataStream.isPublic,
+                isOpenData: newDataStream.isOpenData,
+                isReusable: newDataStream.isReusable,
+                documentationUrl: newDataStream.documentationUrl || undefined,
+                dataLink: newDataStream.dataLink || undefined,
+                unitOfMeasurement: newDataStream.unitOfMeasurement || undefined,
+                dataFrequency: newDataStream.dataFrequency,
+                dataQuality: newDataStream.dataQuality,
+              };
+              await this.sensorService.addDatastream(this.sensor._id, newDataStream);
+            }
+          }
+
+          const currentDataStreams = this.sensor.dataStreams || [];
+          for (const currentDataStream of currentDataStreams) {
+            if (currentDataStream.dataStreamId && !newDataStreamIds.includes(currentDataStream.dataStreamId)) {
+              await this.sensorService.deleteDatastream(this.sensor._id, currentDataStream.dataStreamId);
+            }
+          }
+
+          // TODO: Update existing DataStreams.
+        } catch (error) {
+          console.error(error);
+        }
+
+        // TODO: Execute only if details have changed.
+        try {
+          const details: IUpdateSensorBody = {
+            aim: updatedSensorProperties.aim,
+            description: updatedSensorProperties.description,
+            documentationUrl: updatedSensorProperties.documentationUrl,
+            manufacturer: updatedSensorProperties.manufacturer,
+            name: updatedSensorProperties.name,
+            typeName: updatedSensorProperties.typeName,
+            typeDetails: updatedSensorProperties.typeDetails,
+            theme: updatedSensorProperties.theme,
+          };
+          await this.sensorService.updateDetails(this.sensor._id, details);
+          console.log(`Sensor has successfully been updated.`);
+        } catch (error) {
+          console.error(error);
+        }
+
+        const heightUpdated = this.sensor.location.coordinates[2] !== updatedSensorProperties.location.height;
+        const latitudeUpdated = this.sensor.location.coordinates[1] !== updatedSensorProperties.location.latitude;
+        const longitudeUpdated = this.sensor.location.coordinates[0] !== updatedSensorProperties.location.longitude;
+        if (longitudeUpdated || latitudeUpdated || heightUpdated) {
+          try {
+            await this.sensorService.updateLocation(this.sensor._id, updatedSensorProperties.location);
+          } catch (error) {
+            console.error(error);
+          }
+        }
+
+        await this.router.navigate(['']);
+      } else {
+        const dataStreams = [];
+        const formDataStreams = this.form.value.dataStreams || [];
+        for (const dataStream of formDataStreams) {
+          dataStreams.push({
+            name: dataStream.name,
+            reason: dataStream.reason || undefined,
+            description: dataStream.description || undefined,
+            observedProperty: dataStream.observedProperty || undefined,
+            isPublic: dataStream.isPublic,
+            isOpenData: dataStream.isOpenData,
+            isReusable: dataStream.isReusable,
+            documentationUrl: dataStream.documentationUrl || undefined,
+            dataLink: dataStream.dataLink || undefined,
+            unitOfMeasurement: dataStream.unitOfMeasurement || undefined,
+            dataFrequency: dataStream.dataFrequency,
+            dataQuality: dataStream.dataQuality,
+          });
+        }
+
+        const sensor: IRegisterSensorBody = {
+          dataStreams,
+          typeName: this.form.value.type.typeName,
+          location: this.form.value.location || {},
+          active: JSON.parse(this.form.value.active.value.toLowerCase()) || false, // cast strings ("true") to boolean
+          aim: this.form.value.aim,
+          description: this.form.value.description,
+          documentationUrl: this.form.value.documentationUrl || undefined,
+          manufacturer: this.form.value.manufacturer || undefined,
+          name: this.form.value.name,
+          theme: this.form.value.theme ? this.form.value.theme.value : [],
+          typeDetails: {subType: this.form.value.type.typeDetails || ''},
+        };
+
+        try {
+          await this.sensorService.register(sensor);
+
+          this.locationService.showLocation(null);
+          this.alertService.success('Sensor registered!', false, 4000);
+        } catch (error) {
+          this.alertService.error(`An error has occurred while creating sensor: ${error}.`);
+        }
+      }
+    } else {
+      this.alertService.error(`The form is invalid.`);
+    }
+  }
+
+  public ngOnInit() {
+    const urlRegex = '(https?://)?([\\da-z.-]+)\\.([a-z.]{2,6})[/\\w .-]*/?';
+
+    this.form = this.formBuilder.group({
+      name: ['', [Validators.required, Validators.minLength(6)]],
+      aim: '',
+      description: '',
+      manufacturer: '',
+      active: '',
+      documentationUrl: ['', [Validators.pattern(urlRegex)]],
+      dataStreams: new FormArray([]),
+      location: [],
+      type: [],
+      theme: [],
+    });
+
+    this.formControlSteps = {
+      'Sensor Properties': [
+        this.form.controls.name,
+        this.form.controls.type,
+        this.form.controls.active,
+      ], 'Optional Sensor Properties': [
+        this.form.controls.aim,
+        this.form.controls.description,
+        this.form.controls.manufacturer,
+        this.form.controls.documentationUrl,
+        this.form.controls.theme,
+      ], 'Data Streams': [
+        this.form.controls.dataStreams,
+      ], 'Sensor Location': [
+        this.form.controls.location,
+    ]};
+
+    this.subscriptions.push(
+      this.route.params.subscribe(async params => {
+        if (params.id) {
+          const sensor = await this.sensorService.get(params.id);
+          this.setSensor(sensor as ISensor);
+
+          this.locationService.showLocation(null);
+        }
+      })
+    );
+  }
+
+  public ngOnDestroy(): void {
+    this.subscriptions.forEach((s) => s.unsubscribe());
+  }
+}
