+++ src/app/forms/sensors/sensors.component.html
