+++ src/app/forms/sensors/sensors.component.scss
