+++ src/app/sidebar/sidebar.component.html
@@ -5,17 +5,17 @@
                 <a class="sidebar-link text-center mb-2" routerLink="/owner" routerLinkActive="active" [routerLinkActiveOptions]="{exact:true}">
                     <em class="fas fa-pen fa-lg mx-1"></em>
                     <br>
-                    <em class="side-text mx-1">Edit<br>Owner</em>
+                    <em class="side-text mx-1">Owner</em>
                 </a>
                 <a class="sidebar-link text-center mb-2" routerLink="/sensor" routerLinkActive="active" [routerLinkActiveOptions]="{exact:true}">
                     <em class="fas fa-plus fa-lg mx-1"></em>
                     <br>
-                    <em class="side-text mx-1">Register<br>Sensor</em>
+                    <em class="side-text mx-1">Sensor</em>
                 </a>
                 <a class="sidebar-link text-center mb-2" routerLink="/sensors" routerLinkActive="active" [routerLinkActiveOptions]="{exact:true}">
                     <em class="fas fa-eye mx-1"></em>
                     <br>
-                    <em class="side-text">View<br>Sensors</em>
+                    <em class="side-text">Sensors</em>
                 </a>
                 <a class="sidebar-link text-center" (click)="logout()">
                     <em class="fas fa-sign-out-alt mx-1"></em>
